requirements_dev.txt
python-dotenv
pytest
pyinstaller